chore(lightbox): rename component lightbox
