+++ src/components/style.css
@@ -18,7 +18,7 @@
   position: fixed;
   top: 0px;
   width: 100%;
-  z-index: 2001;
+  z-index: 2000;
   -webkit-align-items: center;
   -moz-box-sizing: border-box;
   -webkit-justify-content: center;
